api/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
SQLAlchemy==2.0.39
alembic==1.15.1
psycopg[binary]==3.2.6
redis==5.2.1
celery==5.4.0
PyJWT==2.10.1
httpx==0.28.1
